[178] move `delimited` inside `p2p/net/wire`
