+++ p2p/net/wire/delimited/copy.go
